+++ package.json
@@ -3,19 +3,19 @@
   "scripts": {
     "bootstrap": "lerna bootstrap",
     "changelog": "lerna-changelog",
-    "content:upload": "npx eubfr-cli content upload",
-    "content:delete": "npx eubfr-cli content delete",
+    "content:upload": "eubfr-cli content upload",
+    "content:delete": "eubfr-cli content delete",
     "deploy": "run-s deploy:resources deploy:services deploy:demo",
-    "deploy:demo": "npx eubfr-cli demo deploy",
+    "deploy:demo": "eubfr-cli demo deploy",
     "deploy:documentation": "./scripts/deployment/deploy-documentation.sh",
     "deploy:once": "./scripts/deployment/deploy-once.sh",
-    "deploy:services": "npx eubfr-cli services deploy",
-    "deploy:resources": "npx eubfr-cli resources deploy",
+    "deploy:services": "eubfr-cli services deploy",
+    "deploy:resources": "eubfr-cli resources deploy",
     "delete": "run-p delete:*",
-    "delete:demo": "npx eubfr-cli demo delete",
+    "delete:demo": "eubfr-cli demo delete",
     "delete:dotfiles": "./scripts/cleaning/delete-dotfiles.sh",
-    "delete:resources": "npx eubfr-cli resources delete",
-    "delete:services": "npx eubfr-cli services delete",
+    "delete:resources": "eubfr-cli resources delete",
+    "delete:services": "eubfr-cli services delete",
     "docs:html": "./scripts/documentation/docs-html.js",
     "docs:md": "./scripts/documentation/docs-md.js",
     "flow": "flow",
