+++ tools/eubfr-cli/commands/services/deleteServices.js
@@ -1,8 +1,3 @@
-const path = require('path');
-const { promisify } = require('util');
-const { exec } = require('child_process');
-
-const shell = promisify(exec);
 const config = require('../../../../config.json'); // eslint-disable-line import/no-unresolved
 
 // Protect certain stages from deletion.
@@ -13,27 +8,13 @@ if (['test', 'acc', 'prod'].includes(config.stage)) {
 
 // Utilities
 const getServices = require('../../lib/getServices');
-const getServiceLocation = require('../../lib/getServiceLocation');
+const deleteServerlessService = require('../../lib/deleteServerlessService');
 
 const deleteServices = async ({ services }) => {
   const list = getServices(services);
 
   const operations = list.map(async service => {
-    const serviceName = service.service;
-    const cwd = path.resolve(`${getServiceLocation(serviceName)}`);
-
-    try {
-      console.log(`Deleting ${serviceName} ...`);
-      console.time(serviceName);
-      await shell(
-        `npx sls remove --stage ${config.stage} --region ${config.region}`,
-        { cwd }
-      );
-      console.log(`${serviceName} has been deleted.`);
-      return console.timeEnd(serviceName);
-    } catch (e) {
-      return console.error(e);
-    }
+    await deleteServerlessService(service.service, {});
   });
 
   return Promise.all(operations);
